Moving UnsafePolicy to config.rs.
